extract common instructions
